+++ packages/grid/src/index.ts
@@ -1,6 +1,8 @@
 export { type BinaryGrid } from './binary-grid';
+export { connectRooms, type Room } from './connect-rooms';
 export { getRegion, getRegions, type Cell } from './flood-fill';
 export { getNeighbours } from './get-neighbours';
+export { jeskoCircle } from './jesko-circle';
 export { randomGrid, type FillOptions, type GridOptions } from './random-grid';
 export { randomWalker, type WalkerOptions } from './random-walker';
 export { smoothGrid } from './smooth-grid';